refactor: 테스트들을 적절한 패키지로 분리해라
