+++ src/test/java/lotto/model/LottoValidatorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
