+++ src/test/java/lotto/service/LottoServiceTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
